refactor: drag and drop 로직 분리
