+++ apps/client/src/features/project/board/components/KanbanBoard.tsx
@@ -1,4 +1,4 @@
-import { ReactNode, useState, DragEvent } from 'react';
+import { ReactNode, DragEvent } from 'react';
 import { Link } from '@tanstack/react-router';
 import { motion, AnimatePresence } from 'framer-motion';
 import { HamburgerMenuIcon, PlusIcon } from '@radix-ui/react-icons';
@@ -27,19 +27,27 @@ import { calculatePosition, findDiff, findTask } from '@/features/project/board/
 import { TaskTextarea } from '@/features/project/board/components/TaskTextarea.tsx';
 import { useBoardStore } from '@/features/project/board/useBoardStore.ts';
 import { SubtaskProgress } from '@/features/project/board/components/SubtaskProgress.tsx';
+import { useDragAndDrop } from '@/features/project/board/useDragAndDrop.ts';
 
 interface KanbanBoardProps {
   projectId: number;
 }
 
 export function KanbanBoard({ projectId }: KanbanBoardProps) {
-  const { sections } = useBoardStore();
-  const { updateTaskPosition, updateTaskTitle, createTask, restoreState } = useBoardStore();
+  const { sections, updateTaskPosition, updateTaskTitle, createTask, restoreState } =
+    useBoardStore();
 
   const toast = useToast();
   const mutations = useBoardMutations(projectId);
-  const [belowSectionId, setBelowSectionId] = useState<number>(-1);
-  const [belowTaskId, setBelowTaskId] = useState<number>(-1);
+
+  const {
+    belowSectionId,
+    belowTaskId,
+    handleDragStart,
+    handleDragOver,
+    handleDragLeave,
+    handleDragEnd,
+  } = useDragAndDrop();
 
   const handleTitleChange = (taskId: number, newTitle: string) => {
     const task = findTask(sections, taskId);
@@ -96,8 +104,7 @@ export function KanbanBoard({ projectId }: KanbanBoardProps) {
     const taskId = Number(event.dataTransfer.getData('taskId'));
 
     if (taskId === belowTaskId) {
-      setBelowTaskId(-1);
-      setBelowSectionId(-1);
+      handleDragEnd();
       return;
     }
 
@@ -113,9 +120,6 @@ export function KanbanBoard({ projectId }: KanbanBoardProps) {
 
     updateTaskPosition(sectionId, taskId, position);
 
-    setBelowTaskId(-1);
-    setBelowSectionId(-1);
-
     mutations.updatePosition.mutate(
       {
         event: 'UPDATE_POSITION',
@@ -130,6 +134,8 @@ export function KanbanBoard({ projectId }: KanbanBoardProps) {
         },
       }
     );
+
+    handleDragEnd();
   };
 
   const handleCreateTask = (sectionId: number) => {
@@ -163,162 +169,129 @@ export function KanbanBoard({ projectId }: KanbanBoardProps) {
     );
   };
 
-  const handleDragStart = (event: DragEvent<HTMLDivElement>, sectionId: number, taskId: number) => {
-    event.dataTransfer.setData('taskId', taskId.toString());
-    event.dataTransfer.setData('sectionId', sectionId.toString());
-  };
-
-  const handleDragEnd = () => {
-    setBelowTaskId(-1);
-    setBelowSectionId(-1);
-  };
-
-  const handleDragOver = (e: DragEvent<HTMLDivElement>, sectionId: number, taskId?: number) => {
-    e.preventDefault();
-    setBelowSectionId(sectionId);
-
-    if (!taskId) {
-      setBelowTaskId(-1);
-      return;
-    }
-
-    setBelowTaskId(taskId);
-  };
-
-  const handleDragLeave = () => {
-    setBelowTaskId(-1);
-    setBelowSectionId(-1);
-  };
-
   return (
-    <div className="relative h-full overflow-hidden">
-      <div className="spazce-x-2 flex h-[calc(100vh-110px)] gap-2 overflow-x-auto p-4">
-        <AnimatePresence mode="popLayout">
-          {sections.map((section) => (
-            <Section
-              key={section.id}
-              className={cn(
-                'flex h-full w-[352px] flex-shrink-0 flex-col items-center',
-                'bg-transparent',
-                section.id === belowSectionId && belowTaskId === -1 && 'border-2 border-blue-400'
-              )}
-            >
-              <SectionHeader className="flex w-full items-center justify-between gap-2 space-y-0">
-                <div className="flex items-center gap-2">
-                  <SectionTitle className="text-xl">{section.name}</SectionTitle>
-                  <SectionCount>{section.tasks.length}</SectionCount>
-                </div>
-                <SectionDropdownMenu>
-                  <Button
-                    type="button"
-                    variant="ghost"
-                    className="text-blac hover:text-primary w-full border-none px-0 hover:bg-white"
-                    onClick={() => handleCreateTask(section.id)}
-                  >
-                    <PlusIcon />
-                    Add Task
-                  </Button>
-                </SectionDropdownMenu>
-              </SectionHeader>
-              <SectionContent
-                key={section.id}
-                className="flex w-full flex-1 flex-col items-center gap-2 overflow-y-auto pt-1"
-                onDragOver={(e) => handleDragOver(e, section.id)}
-                onDragLeave={handleDragLeave}
-                onDrop={(e) => handleDrop(e, section.id)}
-                onDragEnd={handleDragEnd}
-              >
-                {section.tasks.map((task) => (
-                  <motion.div
-                    key={task.id}
-                    layout
-                    layoutId={task.id.toString()}
-                    draggable
-                    initial={{ opacity: 1, zIndex: 1 }}
-                    animate={{
-                      zIndex: task.id === belowTaskId ? 50 : 1,
-                      scale: task.id === belowTaskId ? 1.02 : 1,
-                    }}
-                    transition={{
-                      layout: { duration: 0.3 },
-                      scale: { duration: 0.2 },
-                    }}
-                    style={{ position: 'relative' }}
-                    onDragStart={(e) =>
-                      handleDragStart(
-                        e as unknown as DragEvent<HTMLDivElement>,
-                        section.id,
-                        task.id
-                      )
-                    }
-                    onDrop={(e) => handleDrop(e, section.id)}
-                    onDragOver={(e) => {
-                      e.preventDefault();
-                      e.stopPropagation();
-                      handleDragOver(e, section.id, task.id);
-                    }}
-                    onDragLeave={handleDragLeave}
-                  >
-                    <Card
-                      className={cn(
-                        'w-56 border bg-white transition-all duration-300 md:w-80',
-                        task.id === belowTaskId && 'border-2 border-blue-400',
-                        'hover:shadow-md'
-                      )}
-                    >
-                      <CardHeader className="flex flex-row items-start gap-2 space-y-0">
-                        <TaskTextarea
-                          taskId={task.id}
-                          initialTitle={task.title}
-                          onTitleChange={handleTitleChange}
-                        />
-                        <Button
-                          variant="ghost"
-                          type="button"
-                          asChild
-                          className="hover:text-primary px-2 hover:bg-transparent"
-                        >
-                          <Link to={`/${projectId}/board/${task.id}`} className="p-0">
-                            <PanelLeftOpen className="h-6 w-6" />
-                          </Link>
-                        </Button>
-                      </CardHeader>
-                      <CardContent className="flex items-end justify-between">
-                        <div className="flex flex-wrap gap-1">
-                          {task.labels.map((label) => (
-                            <Badge key={label.id} style={{ backgroundColor: label.color }}>
-                              {label.name}
-                            </Badge>
-                          ))}
-                        </div>
-                        <AssigneeAvatars assignees={task.assignees} />
-                      </CardContent>
-                      {task.subtasks.total > 0 && (
-                        <CardFooter className="flex items-center justify-between space-y-0">
-                          <SubtaskProgress
-                            total={task.subtasks.total}
-                            completed={task.subtasks.completed}
-                          />
-                        </CardFooter>
-                      )}
-                    </Card>
-                  </motion.div>
-                ))}
-              </SectionContent>
-              <SectionFooter className="w-full">
+    <div className="spazce-x-2 flex h-[calc(100vh-110px)] gap-2 overflow-x-auto p-4">
+      <AnimatePresence mode="popLayout">
+        {sections.map((section) => (
+          <Section
+            key={section.id}
+            className={cn(
+              'flex h-full w-[352px] flex-shrink-0 flex-col items-center',
+              'bg-transparent',
+              section.id === belowSectionId && belowTaskId === -1 && 'border-2 border-blue-400'
+            )}
+          >
+            <SectionHeader className="flex w-full items-center justify-between gap-2 space-y-0">
+              <div className="flex items-center gap-2">
+                <SectionTitle className="text-xl">{section.name}</SectionTitle>
+                <SectionCount>{section.tasks.length}</SectionCount>
+              </div>
+              <SectionDropdownMenu>
                 <Button
+                  type="button"
                   variant="ghost"
-                  className="w-full border-none px-0 text-black"
+                  className="text-blac hover:text-primary w-full border-none px-0 hover:bg-white"
                   onClick={() => handleCreateTask(section.id)}
                 >
                   <PlusIcon />
                   Add Task
                 </Button>
-              </SectionFooter>
-            </Section>
-          ))}
-        </AnimatePresence>
-      </div>
+              </SectionDropdownMenu>
+            </SectionHeader>
+            <SectionContent
+              key={section.id}
+              className="flex w-full flex-1 flex-col items-center gap-2 overflow-y-auto pt-1"
+              onDragOver={(e) => handleDragOver(e, section.id)}
+              onDragLeave={handleDragLeave}
+              onDrop={(e) => handleDrop(e, section.id)}
+              onDragEnd={handleDragEnd}
+            >
+              {section.tasks.map((task) => (
+                <motion.div
+                  key={task.id}
+                  layout
+                  layoutId={task.id.toString()}
+                  draggable
+                  initial={{ opacity: 1, zIndex: 1 }}
+                  animate={{
+                    zIndex: task.id === belowTaskId ? 50 : 1,
+                    scale: task.id === belowTaskId ? 1.02 : 1,
+                  }}
+                  transition={{
+                    layout: { duration: 0.3 },
+                    scale: { duration: 0.2 },
+                  }}
+                  style={{ position: 'relative' }}
+                  onDragStart={(e) =>
+                    handleDragStart(e as unknown as DragEvent, section.id, task.id)
+                  }
+                  onDrop={(e) => handleDrop(e, section.id)}
+                  onDragOver={(e) => {
+                    e.preventDefault();
+                    e.stopPropagation();
+                    handleDragOver(e, section.id, task.id);
+                  }}
+                  onDragLeave={handleDragLeave}
+                >
+                  <Card
+                    className={cn(
+                      'w-56 border bg-white transition-all duration-300 md:w-80',
+                      task.id === belowTaskId && 'border-2 border-blue-400',
+                      'hover:shadow-md'
+                    )}
+                  >
+                    <CardHeader className="flex flex-row items-start gap-2 space-y-0">
+                      <TaskTextarea
+                        taskId={task.id}
+                        initialTitle={task.title}
+                        onTitleChange={handleTitleChange}
+                      />
+                      <Button
+                        variant="ghost"
+                        type="button"
+                        asChild
+                        className="hover:text-primary px-2 hover:bg-transparent"
+                      >
+                        <Link to={`/${projectId}/board/${task.id}`} className="p-0">
+                          <PanelLeftOpen className="h-6 w-6" />
+                        </Link>
+                      </Button>
+                    </CardHeader>
+                    <CardContent className="flex items-end justify-between">
+                      <div className="flex flex-wrap gap-1">
+                        {task.labels.map((label) => (
+                          <Badge key={label.id} style={{ backgroundColor: label.color }}>
+                            {label.name}
+                          </Badge>
+                        ))}
+                      </div>
+                      <AssigneeAvatars assignees={task.assignees} />
+                    </CardContent>
+                    {task.subtasks.total > 0 && (
+                      <CardFooter className="flex items-center justify-between space-y-0">
+                        <SubtaskProgress
+                          total={task.subtasks.total}
+                          completed={task.subtasks.completed}
+                        />
+                      </CardFooter>
+                    )}
+                  </Card>
+                </motion.div>
+              ))}
+            </SectionContent>
+            <SectionFooter className="w-full">
+              <Button
+                variant="ghost"
+                className="w-full border-none px-0 text-black"
+                onClick={() => handleCreateTask(section.id)}
+              >
+                <PlusIcon />
+                Add Task
+              </Button>
+            </SectionFooter>
+          </Section>
+        ))}
+      </AnimatePresence>
     </div>
   );
 }
